Reorganize and update boilerplate to latest recs
Based on the latest tomo-plugin template:

- Add rubocop-minitest
- Loosen up rubocop guidelines
- Point to new CircleCI pipelines UI
- Add TOC, and support section to the README
- Add `bin/console` and `bin/setup` scripts
- Declare `Tomo::Plugin::Rollbar` instead of using nested module syntax
- Move task tests into `rollbar/tasks_test.rb`
- Pin tomo dependency at `~> 1.0`
