+++ .gitignore
@@ -4,6 +4,7 @@
 /coverage/
 /doc/
 /pkg/
+/site/
 /spec/reports/
 /tmp/
 /Gemfile.lock
