+++ .rubocop.yml
@@ -1,4 +1,6 @@
-require: rubocop-performance
+require:
+  - rubocop-minitest
+  - rubocop-performance
 
 AllCops:
   DisplayCopNames: true
@@ -8,10 +10,6 @@ AllCops:
     - "tmp/**/*"
     - "vendor/**/*"
 
-Layout/LineLength:
-  Exclude:
-    - "*.gemspec"
-
 Layout/HashAlignment:
   EnforcedColonStyle:
     - table
@@ -19,6 +17,10 @@ Layout/HashAlignment:
   EnforcedHashRocketStyle:
     - table
     - key
+
+Layout/LineLength:
+  Exclude:
+    - "*.gemspec"
 
 Layout/SpaceAroundEqualsInParameterDefault:
   EnforcedStyle: no_space
@@ -30,7 +32,7 @@ Lint/StructNewOverride:
   Enabled: true
 
 Metrics/AbcSize:
-  Max: 16
+  Max: 20
   Exclude:
     - "test/**/*"
 
@@ -44,8 +46,12 @@ Metrics/ClassLength:
     - "test/**/*"
 
 Metrics/MethodLength:
+  Max: 18
   Exclude:
     - "test/**/*"
+
+Metrics/ParameterLists:
+  Max: 6
 
 Naming/MemoizedInstanceVariableName:
   Enabled: false
@@ -68,6 +74,9 @@ Style/DoubleNegation:
 Style/EmptyMethod:
   Enabled: false
 
+Style/FormatStringToken:
+  Enabled: false
+
 Style/FrozenStringLiteralComment:
   Enabled: false
 
@@ -82,6 +91,9 @@ Style/HashTransformKeys:
 
 Style/HashTransformValues:
   Enabled: true
+
+Style/NumericPredicate:
+  Enabled: false
 
 Style/StringLiterals:
   EnforcedStyle: double_quotes
