+++ bin/console
@@ -1,0 +1,15 @@
+#!/usr/bin/env ruby
+
+gem "bundler", "~> 2.0"
+require "bundler/setup"
+require "tomo/plugin/rollbar"
+
+# You can add fixtures and/or initialization code here to make experimenting
+# with your gem easier. You can also use a different console, if you like.
+
+# (If you use this, don't forget to add pry to your Gemfile!)
+# require "pry"
+# Pry.start
+
+require "irb"
+IRB.start(__FILE__)
