+++ lib/tomo/plugin/rollbar.rb
@@ -2,13 +2,11 @@ require "tomo"
 require_relative "rollbar/tasks"
 require_relative "rollbar/version"
 
-module Tomo::Plugin
-  module Rollbar
-    extend Tomo::PluginDSL
+module Tomo::Plugin::Rollbar
+  extend Tomo::PluginDSL
 
-    tasks Tomo::Plugin::Rollbar::Tasks
+  tasks Tomo::Plugin::Rollbar::Tasks
 
-    defaults rollbar_env:   nil,
-             rollbar_token: nil
-  end
+  defaults rollbar_env:   nil,
+           rollbar_token: nil
 end
