+++ lib/tomo/plugin/rollbar/version.rb
@@ -1,7 +1,8 @@
 module Tomo
   module Plugin
-    module Rollbar
-      VERSION = "0.1.1".freeze
-    end
   end
+end
+
+module Tomo::Plugin::Rollbar
+  VERSION = "0.1.1".freeze
 end
