+++ test/test_helper.rb
@@ -1,4 +1,6 @@
 $LOAD_PATH.unshift File.expand_path("../lib", __dir__)
+require "tomo/plugin/rollbar"
 require "tomo/testing"
+
 require "minitest/autorun"
 Dir[File.expand_path("support/**/*.rb", __dir__)].sort.each { |rb| require(rb) }
