+++ tomo-plugin-rollbar.gemspec
@@ -20,17 +20,14 @@ Gem::Specification.new do |spec|
   }
 
   # Specify which files should be added to the gem when it is released.
-  # The `git ls-files -z` loads the files in the RubyGem that have been added into git.
-  spec.files = Dir.chdir(File.expand_path(__dir__)) do
-    `git ls-files -z`.split("\x0").reject { |f| f.match(%r{^(docs|test|spec|features)/}) }
-  end
+  spec.files = `git ls-files -z exe lib LICENSE.txt README.md`.split("\x0")
   spec.bindir        = "exe"
   spec.executables   = spec.files.grep(%r{^exe/}) { |f| File.basename(f) }
   spec.require_paths = ["lib"]
 
   spec.required_ruby_version = ">= 2.5.0"
 
-  spec.add_dependency "tomo"
+  spec.add_dependency "tomo", "~> 1.0"
 
   spec.add_development_dependency "bundler", "~> 2.0"
   spec.add_development_dependency "minitest", "~> 5.11"
@@ -39,6 +36,7 @@ Gem::Specification.new do |spec|
   spec.add_development_dependency "rake", "~> 13.0"
   spec.add_development_dependency "rexml", "~> 3.2"
   spec.add_development_dependency "rubocop", "0.81.0"
+  spec.add_development_dependency "rubocop-minitest", "0.8.1"
   spec.add_development_dependency "rubocop-performance", "1.5.2"
   spec.add_development_dependency "webmock", "~> 3.6"
 end
